% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzdatar_utilities.R
\name{get_data_library_path}
\alias{get_data_library_path}
\title{Retrieve the Data Library Path for nzdatar}
\usage{
get_data_library_path()
}
\value{
A character string representing the path to the data library.
If the path has not been set, this function will stop and prompt the user to set it.
}
\description{
This function retrieves the \code{NZDATAR_LIBRARY_PATH} set in the current R session's options,
which should have been initialized from the user's \code{.Renviron} file upon the start of the session.
It is essential that the path is set before using any functionalities that rely on this setting.
}
\examples{
\dontrun{
library_path <- get_data_library_path()
}
}
