% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzdatar-package.R
\docType{package}
\name{nzdatar-package}
\alias{nzdatar}
\alias{nzdatar-package}
\title{nzdatar: Provides up-to-date data for New Zealand}
\description{
The purpose of this package is to provide a series of functions that connect to a nzdatar library. This library needs to be downloaded separately.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://phildonovan.github.io/nzdatar/}
}

}
\author{
\strong{Maintainer}: Phil Donovan \email{phildonovan99@protonmail.com} (\href{https://orcid.org/0009-0008-7600-8224}{ORCID})

}
\keyword{internal}
