% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzdatar_utilities.R
\name{set_data_library_path}
\alias{set_data_library_path}
\title{Set the Data Library Path for nzdatar}
\usage{
set_data_library_path(path)
}
\arguments{
\item{path}{The file path to the data library that you want to set.}
}
\value{
Invisible NULL. The function is used for its side effect of writing to the \code{.Renviron}.
}
\description{
This function sets the \code{NZDATAR_LIBRARY_PATH} in the user's \code{.Renviron} file,
making it persistent across R sessions. If the \code{.Renviron} file does not exist,
it will be created. Use this function to configure the path to your data library
before using the nzdatar package functionalities.
}
\examples{
\dontrun{
set_data_library_path("/path/to/your/data_library")
}
}
